IAT-2174: Items with only a single master branch are flagged as failing within item report due to being in a state of MERGE.
Think an error is getting swallowed.  Nothing is showing in the logs and there is indeed a place where an exception is not logged at all.  It appears the recent fix where we delete the local repo and clone it fresh from the remote origin is failing.  Added proper logging for the exception.